+++ src/main/java/com/example/dorywcza/repository/ServiceDataRangeRepository.java
@@ -3,5 +3,5 @@ package com.example.dorywcza.repository;
 import com.example.dorywcza.model.service_offer.ServiceDateRange;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ServiceDataRangeRepository extends JpaRepository<ServiceDateRange, Integer> {
+public interface ServiceDataRangeRepository extends JpaRepository<ServiceDateRange, Long> {
 }
