+++ src/main/java/com/example/dorywcza/repository/ServiceJobSalaryRepository.java
@@ -3,5 +3,5 @@ package com.example.dorywcza.repository;
 import com.example.dorywcza.model.service_offer.ServiceJobSalary;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ServiceJobSalaryRepository extends JpaRepository<ServiceJobSalary, Integer> {
+public interface ServiceJobSalaryRepository extends JpaRepository<ServiceJobSalary, Long> {
 }
