+++ src/main/java/com/example/dorywcza/repository/ServiceLocationRepository.java
@@ -3,5 +3,5 @@ package com.example.dorywcza.repository;
 import com.example.dorywcza.model.service_offer.ServiceLocation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ServiceLocationRepository extends JpaRepository<ServiceLocation, Integer> {
+public interface ServiceLocationRepository extends JpaRepository<ServiceLocation, Long> {
 }
