+++ src/main/java/com/example/dorywcza/repository/ServiceScheduleRepository.java
@@ -3,5 +3,5 @@ package com.example.dorywcza.repository;
 import com.example.dorywcza.model.service_offer.ServiceSchedule;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ServiceScheduleRepository extends JpaRepository<ServiceSchedule, Integer> {
+public interface ServiceScheduleRepository extends JpaRepository<ServiceSchedule, Long> {
 }
